some support for differentiating between known and unknown-length arrays (though code-gen hasn't yet changed).
also added ‘--‘ back into parser

+++ src/AST/generate_code.cpp
@@ -35,7 +35,7 @@ extern llvm::Value *null(Type *t);
 extern llvm::ConstantInt *const_true();
 extern llvm::ConstantInt *const_false();
 extern llvm::ConstantInt *const_uint(size_t n);
-extern llvm::ConstantInt *const_int(int n);
+extern llvm::ConstantInt *const_int(long n);
 extern llvm::ConstantInt *const_char(char c);
 extern llvm::ConstantFP *const_real(double d);
 extern llvm::Value *global_string(const std::string &s);
